Test to optimize tox
- We don't need to delete the whole schema for each test.

The schema is only deleted for destructive tests (like migration) and
for all other tests we only truncate all the tables.

- We also introduce a new function to reset the file event increment.

Now that the schema is not delete at each step the autoincrement is
never reseted and the tests are not valid. So with this reset tests
stay green

- We add directory where py.test has nothing to do

Even if directory are not huge it help py.test to go faster

Change-Id: I4770601df0ced2b52f711b0a4dbe2b4cdbdfc715