schema/job: don't enforce ID on the value
The value sent in remoteci_id is not used anymore, hence its migration
to Optional in a previous review. It hasn't been removed because we need
to keep backward compatibility with clients that did not update their
package locally yet.

Since we don't rely on this value, we shouldn't ensure that it is a
UUID, it can be of any type.

Change-Id: Ifcf90efc663fd00b6b930b9a388457de14f3cd93