Use authenticated email account for sending notifications
Up until now, the control-server was relying on the host to act as a
relay to send notifications. This involved no configuration, simply
sending the email to the local postfix and ensuring postfix as a relay
configured.

We did not have the control of the relay server postfix was using.
Now Distributed-CI has a mail server configured where we can control
SPF/DKIM and other security related configuration. This change aims
for the worker to use an authenticated account to send emails.

Change-Id: I2d9a24ab4e07714367b6be0f078b0d04a31c1c61