[regressions]: add columns to tests_results
This commit adds two new column to the tests_results table:
    - regressions: the number of regressions
    - testscases: the testscases themselves

Also, it adds the function add_regressions_and_successfix_to_tests
in transformation.py

This funcion takes two dicts testssuites and compute the
regressions and the successfix.

The regression is set in the following:
    - when a testcase switch from success to failure
    - when a testcase was in regression and still in failure

A successfix testcase correspond to the switch from failure
to passed state.

Change-Id: I528444649865cb18993ec72809561aae8977f09a