Update patch set 17
Patch Set 17:

Open question: should we display all the team attached to a topic even if they don't have a run yet, or do we display only the team that made a run.

From a Product Manager point of view, if only 20% of my partner has run the job, 100% success doesn't mean 100% of my partners are in success. So not saying the 100% in the green block should be changed, but at least have a visual hint of the list of the partner that haven't run it.

Patch-set: 17