Update patch set 6
Patch Set 6: Code-Review-1

dci/api/v1/scheduler.py look like a set of utils functions all doing some DB calls used in the job scheduler. But it will be better if all related DB call are placed in the good file. get_last_components() in the components file, get_remoteci_configuration in remotecis file. etc

We will discover some duplication. Some methods already exists in utils or in v1_utils. It will be better to reuse it.

We can also work together to avoid coupling business domain to the SQL. The goal is to have a codebase that is loosely coupled and high cohesive, so that changes are easy, fast and safe to make.

Patch-set: 6
Label: Code-Review=-1