Introduce the error state
This commit introduces the error state as a jobstate. A job would result
in error if anything wrong happens during the setup of the platform (new
and pre-run). This would mean that the job isn't usefull for
metrics/feedback as it didn't get to test the actual product.

Change-Id: I77ad56ea8f83b8ee7e129864d09b66cd81e90044
Trello: https://trello.com/c/FK6v7k35