Replace next_topic by next_topic_id
The choice of the variable name next_topic is not consistent with the
way the variables are generally named in our code base, leading to
specialized code to handle this.

This commit aims to rename this variable in a way that is consistent
with our naming scheme, since it represents an id, it is called
next_topic_id.

The change will be done in two steps, first one - this patch - will
manage backward compatibility, once the various library will be updated
to rely on next_topic_id, next_topic will be completely removed.

Change-Id: Icb27a657f437ebb5b1a981fe171ebd46874e1153