Ensure [_meta][count] is always accurate
Currently, when using limit in the URL, the value in [_meta][count] is
incorrect. It returns the count of all the values, irrespectively of the
limit argument.

This is because utils.get_number_row_count() isn't taking limit in
consideration. The method itself isn't necessary as we can rely directly
on len() to return the number of rows, hence the method has been deleted
and all the call to get_number_row_count() replaced by len().

Change-Id: I99239fca159773c5207ea71b6df644decd0fb223
Fixes: https://github.com/redhat-cip/dci-control-server/issues/57