Update patch set 4
Patch Set 4: Code-Review-1

That's a good thing to accelerate the db-prov script, but for me manually adding object with uuid, date etc ... into a python script is like manually managing a SQLDump. It will be hard to change when model change, it will be hard to add object in it, etc ... On optimisation can be to just insert a sqldump and add an option to this script to save the running model/data to this same SQLDump or load the saved dump.

So the process will be easier to maintain a test dataset :
- We load the dump (apply the model change if needed)
- Play with the API/UI to add data/modify them
- We save the new dataset as a dump
- We commit the dump as the new ref dataset of test

Patch-set: 4
Reviewer: Gerrit User 107 <107@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1