job.create: honor the previous_job_id parameter
Add the ability to create a sub-job based on an existing one.

The `previous_job_id` key is used to create update or upgrade jobs.
Until now, the key was only used by the /update and /upgrade end-points.
This commit adds it to the generic /jobs POST end-point. This allow the
creation of more specific upgrade job, like the RHOSP FFU upgrade.

Change-Id: Ifaf6b507f2b114bdc132868de1e4f579f6d0afaa