Fix pip version to match rpm version
Remove Werkzeug because it's a dependency of flask.

Use the version of alembic we have in epel (0.8.3).
The version we have in centos (0.4.2) doesn't work.

We install the latest version of psycopg2, cryptography, pyopenssl and lxml
because we cannot build lib during the pip install

Remove test in spec file because no test are run

Remove pylint because it's not use anymore

Change-Id: Ia5d151b5e7c1e80cf69bfae028d47571f7276b5d