Remove tests resource usage
Tests that are tied to Jobs, Teams, and Remotecis
have never been used before. This patch get rid
of this feature.
It also prepare the feature to associate
individual tests to issues.

TESTS.team_id has been deleted from the database
schema but it's still around the API in order
to not break dciclient tests.

Change-Id: I287bb37a65e74a275221f20e0937553a1c843c4f