global_status.py: Filter components on state
Currently when a new component failed to be synced the state remains
inactive. Because it's the latest component, the global status will
try to find a job associated but it's not possible.
As a solution we need to filter the component state with only the
active value.
Also reformat the SQL sub query.

Change-Id: Ib834832806a2db86f6f35a0984d5311d340d1d4c