stores: Fix the file engine upload method
When using the filesystem store engine, we are manipulating a string
and not a stream during upload (due to request.data). So there's not
read attribute to iterable variable so we just need to write the
complete string to the file.
Also set a more appropriate content-type to head method.

Change-Id: I7d4d8b7a79e76daf315fc0b47b7731025101b484
Fixes: #74