stores: Fix the file engine upload method
When using the filesystem store engine, we are manipulating either a
string (request.data) or a stream (request.stream) during upload.
So there's not read attribute on iterable variable when we're in the
case of using request.data so we just need to write the complete
string to the file.
Also set a more appropriate content-type to head method.

Change-Id: I7d4d8b7a79e76daf315fc0b47b7731025101b484
Fixes: #74