Update patch set 19
Patch Set 19:

Hello Yassine,
in the identity object I need the email, fullname, and timezone of the current user (see /users/me endpoint).

The team returned in identity.teams should be also identical the the teams endpoint. We are missing the team id. I need also the field 'team_name' to be 'name'.

Patch-set: 19