Update patch set 6
Patch Set 6:

global policies.py file won't help yassine.
When you do     

    if (user.is_not_super_admin() and user.is_not_read_only_user() and
        user.is_not_epm()):
        query.add_extra_condition(_TABLE.c.team_id.in_(user.teams_ids))

You add extra condition depending on the role, you are not returning a 401 forbidden.

Here we would like to know if the user is a simple user, positive if are always best because you are not oblige to update your code when you add a new role. You only check that the user.is_simple_user() is false with the new role.

Patch-set: 6
Label: Code-Review=0