Update patch set 3
Patch Set 3: Code-Review-1

I don't think its a good idea to give the EPM such permissions.

A user should be promoted as an EPM by the DCI admin team and not by other EPMS.

A Red Hat user is automatically associated if it has the required JWT field, if we have to add a user to the Red Hat team this will be very rare and exceptional thus must be done by a DCI admin and not by an EPM.

An EPM cannot add himself to the admin team, i really don't think this  is a good idea...

Special teams must be managed carefully only by the DCI admin team.

Patch-set: 3
Label: Code-Review=-1