Update patch set 8
Patch Set 8: Code-Review+1

This patch introduce more requests on the cert endpoint. Typically for a RHEL compose we do 5k requests per download.

And I don't like this granularity in the "verification" process. For me if a user try to GET a resource, the server should return 404 if the resource is state inactive or doesn't exists.

It would be great to have one method validating if an identity can access a topic. We use this in a lot of place (get and list in components and topics).

I found this is hard to reuse somewhere else in the code.
You understand it, I'm not a big fan of this one, it fixes an important bug so +1 but it would be great to add some refactoring in the code after.

Patch-set: 8
Reviewer: Gerrit User 405 <405@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1