[products/topics/teams]: fix permission consistency
Once a team is deleted through

  DELETE /teams/<team-id>

The team is actually still in the DB, only the
state switched to 'archived'. Thus when we get
the teams associated to a product or a topic,
we still see the previous deleted team.

This patch fix this issue by adding an
extra condition when getting the permission.

Change-Id: I6a6997c7a41ac9e7f00bbc04504dacb4576808d1