Update patch set 2
Patch Set 2:

After brainstorming session this morning we ended with:

A team should have the flag "can_access_topic_without_earf" if the topic doesn't have an "earf_signed" flag to be true. EARF is for Export Authorization Request Form. It's a legal term.

so in this patch we can introduce team.can_access_topic_without_earf ?

In an other patch we can change topic.export_control to topic.earf_signed

Patch-set: 2