remove login_required for /certs/verify endpoint
This endpoint is used by Nginx/repo.distributed-ci.io to decide if a
user identified by an SSL certificate is allowed to access files inside
a certain path (/:product_id/:topic_id/:component_id).

It currently allows users with valid credentials, Nginx uses basic auth
for this and our basic auth credentials verification is quite CPU
intensive (~ 500 ms) & the remaining of the request is generally 10 to
50 times faster (10 ms < ~ < 50 ms).

As we're trying to enable concurrency in dci-downloader to improve
download times for the partners, this is becoming a problem as it leads
to the server seeing 100% CPU usage with 4 concurrent requests and a
maximum of 100~120 requests/minute, slowing down all other API
operations.

Given that:

- all parameters are either random UUIDs or a SSL certificate
  fingerprint (can be considered as random) and cannot be guessed,

- even guessing this data would bring no value to a potential attacker,

- the endpoint returns only "200 - OK" or "403 - Explanation" and no other
  content,

- the certificate fingerprint is matched against the remoteCIs list from
  the database,

- the path elements are matched against product/topic/component ids from
  the database.

It seems useless to restrict this API endpoint with any login/password
or source IP address.

Change-Id: I58fbad582a2e959df9dc6306cb1c8954d2dfaecf