Set more realistic sqlalchemy pool size
According to https://docs.sqlalchemy.org/en/13/core/pooling.html

pool_size is defined as:
"This is the largest number of connections that will be kept
persistently in the pool. Note that the pool begins with no connections;
once this number of connections is requested, that number of connections
will remain"

max_overflow is defined as:
"When the number of checked-out connections reaches the size set in
pool_size, additional connections will be returned up to this limit.
When those additional connections are returned to the pool, they are
disconnected and discarded. It follows then that the total number of
simultaneous connections the pool will allow is pool_size +
max_overflow, and the total number of “sleeping” connections the pool
will allow is pool_size. "

5 stale connections should be plenty enough for our needs.
20 connections total should be enough also.

(this is per httpd process which is by default 3 < processes < 16)

Change-Id: I9121aa1555153340a28cf34172c1f62351b817b5