Update patch set 4
Patch Set 4:

This has not been defined yet, it could be either from the UI or from a script i guess. What i can say now is that it will not be part of the feeder.

Btw, using the remoteci to create a components is not a "restricted" stuff because every user can get the credentials of every remotecis of its team.

Patch-set: 4