Update patch set 2
Patch Set 2: Code-Review-1

Hello,

If I have to choose, I would like to preserve the consistency between resources. Meaning created_at (auto field for resource creation) and modified_at (when the resource was modified).

If I'm correct here you need a way to sort component by date. What will be the good name for this date ?

If this date correspond to the moment when the component was released, maybe released_at is a good name.

So I would prefer also to use an extra variable instead of an exception for a resource

Patch-set: 2
Label: Code-Review=-1