Update patch set 5
Patch Set 5: Code-Review-1

Yassine, the issue is updating a resource we previously get doesn't work anymore.

If a user do a put or post with an extra field like "ImAHacker", I want to keep the 400 error saying "ImAHacker is not a valid field".

But if I get a user, the api return an object with "team" field. I would like the api to accept "team" field in a put or a post. Not raise a 500 error.

So in our code, we have 2 solutions:

1/ we remove the extra related field "team" on the resource
2/ we make the ORM to accept related field and ignore them

2/ is probably better because it's generic

Do you see another solution ?

Patch-set: 5
Reviewer: Gerrit User 405 <405@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1