Update requirements.txt and test-requirements.txt
Versions in requirements.txt are fixed to be similar to the production.
The issue is that we don't update the version when a package is updated.

So we can remove the version in the test-requirements.txt and use the
latest version in the package.

We will discuss in the future to see if we need to remove the fixed
version in the requirements.txt

Change-Id: Idf3522284ad3fc039f60c8e6e07def343dd9d7ab