Update requirements.txt and test-requirements.txt
Versions in requirements.txt are fixed to be similar to the production.
The issue is that we don't update the version when a package is updated
on production.

This patch remove the fixed version in the test-requirements.txt and use the
latest version in the package. Those packages are not used in the
production server, only in the test.

We will discuss in the future to see if we need to remove the fixed
version in the requirements.txt. But in the mean time, we update the
version to match the one we have on the production server.

Change-Id: Idf3522284ad3fc039f60c8e6e07def343dd9d7ab