Fix breaking change with Py3 & mod_wsgi.Input
To avoid copying the data in-memory or locally we used to try to pass
`flask.request.stream` instead of `flask.request.data` when receiving
files (specially when using Swift backend).

It appears that mod_wsgi never supported the use of `.tell()` [1] and
this maybe has never worked as we thought on the production environment
at least.

In addition since we're using DCIAuth as authentication method 99.9% of
the time, the stream is already consumed anyway during the content
checksum, making this optimization attempt useless. [2]

This change removes the `upload_component_file()` function and replaces
its calls with `flask.request.data`.

[1]: https://github.com/GrahamDumpleton/mod_wsgi/issues/545#issuecomment-601109281
[2]: from production logs, in August 2021:
  - `request.stream` was returned 2 times
  - `request.data` was returned 743300 times

Change-Id: I4797c18498237a297351fd4c40a8de69a9780180