Fix add_regressions_and_successfix_to_tests to avoid changing the name of some testcases
I don't know why but we split the test name using a bracket and keep only the first part.
We put this variable in the testcase name changing is value.

If testcase name contains a '[' then the name is changing good.
This patch remove the side effect and let the name untouched.

Change-Id: Icfb3969495f20a2af219780bee5051e172d31ff0