Fix error 500 when topic is not active
When trying to access a file for an inactive topic, we get the following
`TypeError: 'Topic' object is not subscriptable`.

With recent move to ORM, database fetches into objects rather than in
arrays. There was a leftover for topic which was only evaluated when
topic is inactive.

Change `topic[…]` with `topic.…`

Change-Id: I8a55b02072b6b42cb4655ca7e56d7fbc4e8b38d3