components/s3: generate aws S3 presigned urls
DCI components will be available on S3. We can generate a pre signed url to get access to those components.
In order to prevent transversal attack, we normalize the relative path into an absolute path and compare to the relative path.
Return a 400 malformed request if path is not good.

Change-Id: Iedc2874386d26fe4715f11c8d40351f5834f0313