Change default string field in job model to empty string
If we create a job, get it with the API and update it as this, without any modification, it's not working.
This patch fixes the issue and change the default value of text and string field in jobs to empty string
The empty string "" is use for the default value.

Change-Id: I8323ecec6076dbbbf71414d64d548b53a77b9bc6