Force no-binary for psycopg2
On python3.6, pip installs psycopg2==2.7.5 from a prebuilt wheel which
brings in its own libpq and libssl. This libssl doesn't work with
python's libssl (from the system) and prevents SSL connections (core
dump) to PostgreSQL servers. Cf. [1]

Adding `--no-binary=psycopg2` allows to force to build from source.
Additionally RHEL's libpq needs to be kept installed on the machine.

Also, `--no-binary=psycopg2` could be added to requirements.txt but pip
fails with it when installing with `--editable .` for some reason.

[1] https://github.com/psycopg/psycopg2/issues/543#issuecomment-295015478

Change-Id: I6e1c3d5d0659cf59b97b9d2562b6c467f2c2f723