Make the dci-dbinit script behave correctly on prod-like env
On pro-like env, the dci user can't access the postgresql database or
create a database. However the database exists.

Just wrap the database check/creation into an exception handler to
ignore failure to check existence or create the database.

If the database doesn't exist and can't be created, the script will
still fail later.

Change-Id: I608cf3f5deab02e76de7fc013a5c736f80658d7d